requirements.txt
pygame==2.5.2
opencv-python==4.8.1.78
mediapipe==0.10.13
google-generativeai==0.3.0
pyttsx3==2.90
python-dotenv==1.0.0
pyautogui==0.9.54
pywin32==306
flask==3.0.0
flask-cors==4.0.0
pyperclip==1.8.2
